refactor: update import paths to use explicit file extensions and improve component naming consistency
