+++ src/domains/auth/components/button/submit-button.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/button.tsx";
 import clsx from "clsx";
 import { Loader2 } from "lucide-react";
